show stat: average first name length
Wanted to know how to mock and test promises and error

What I need to know:

a. toThrowError is used for comparing error message
b. toThrow is used for comparing error

c. if (!array.length) can be used for check empty array. Don't
be confused with !array.

d. throw new Error not throw Error because Error is class so it
needs to be instantiated before throwing it.

e. reduce can be used for sum with the following:
[1, 2, 3].reduce((a, b) => a + b, 0);

It is much more similar to that of python.